+++ client/src/redux/reducers/index.js
@@ -54,6 +54,7 @@ const initialState = {
   isLoading: true,
   usersLoged: [],
   brands: [],
+  brandsCopy: [],
   products: [],
   searchProduct: [],
   productsSort: [],
@@ -126,6 +127,7 @@ export default function rootReducer(state = initialState, { type, payload }) {
       return {
         ...state,
         brands: brandFilter,
+        brandsCopy: brandFilter,
       };
     case FILTER_BY_BRAND:
       if (payload === "all") {
@@ -435,7 +437,7 @@ export default function rootReducer(state = initialState, { type, payload }) {
       };
 
     case DEL_FAV:
-      return { ...state};
+      return { ...state };
 
     case ADD_DIRECCIONES:
       return {
@@ -485,5 +487,3 @@ export default function rootReducer(state = initialState, { type, payload }) {
       return state;
   }
 }
-
-
